crux-util/src/test/java/com/fizzed/crux/util/WaitForTest.java
/*
 * Copyright 2017 Fizzed, Inc.
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *      http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */
package com.fizzed.crux.util;

import java.util.concurrent.TimeUnit;
import static org.hamcrest.CoreMatchers.is;
import static org.hamcrest.MatcherAssert.assertThat;
import org.junit.Test;

public class WaitForTest {
    
    @Test
    public void await() throws InterruptedException {
        assertThat(WaitFor.of(() -> true).await(1000000, 10, TimeUnit.MILLISECONDS), is(true));
        assertThat(WaitFor.of(() -> false).await(10, 10, TimeUnit.MILLISECONDS), is(false));
    }
    
}
